11/22/23: use fixtures when seeding
Since we already have some basic fixture data, we can use that
to create our development data. That way we don't have to do work
in both the fixtures and db/seeds.rb, and we can keep test data
and development data in sync. In this commit, we make the
'rails db:seed' command the same as 'rails db:fixtures:load'. This
essentially will remove all the quotes in the db, and load  the
fixtures as dev data.