11/27/23: purposely add red tests
I want to utilize Turbo to have the "new" and "edit" functionality
all render on the #index page, vs. having to go to the actual #new
and #edit pages. So, I'll start by refactoring these tests and
making them purposely fail; the goal is to get them to pass as green
by implementing turbo frames.