11/27/23: add comment about first turbo frame
I added a helpful comment to explain more what the
turbo_frame_tag does. However, if I were to create a new quote
as is, it renders a "Content missing" text on the page. I believe
this is due to the fact that Turbo is expecting a frame of
the same id on the target page in order to replace the frame's
content on the source page with the frame's content on the target
page. Since the #new partial doesn't have that frame with the
same id, it is triggering the error.