11/27/23: add turbo frame for new quote on index page
I left comments with more context on the partials themselves,
but essentially the "New quote" link on the index page will have
a data_turbo_frame attribute that points to a specific turbo frame
within the #new partial. Once the server responds to the AJAX
request with the HTML, the server will add the specific frame's
HTML to the target frame on the #index page.