11/27/23: prepend newly created quote
This commit adds a new turbo_frame with an id of "quotes" on the
that will prepend the newly created quote to the list of existing
quotes (or whatever's in that "quotes" turbo_frame. Turbo will also
make sure to remove the form to add a new quote after a quote is
submitted via the form.