11/27/23: add Cancel functionality
This commit adds a minimal amount of code, because Turbo Frames
already does most of the heavy lifting for us. Essentially, adding
the Cancel button on the _form.html.erb partial enables us to
redirect seamlessly to the #index page and remove the form. If
you were on the #edit page, then the original HTML for that quote
will replace the form. If you were on the #new page, then the
empty frame will replace the form.