11/28/23: add users via Devise
This commit generates a lot of the boilerplate logic for adding
users to the application with the popular Devise gem. With these
Devise users in place, users will now be able to login and view
the quotes associated with their companies. Since the primary
purpose of this tutorial is to learn Turbo Streams and security,
we are not interested in coding the logic to attach new users
to their companies. Perhaps in a future commit I will allow this.