11/28/23: render current_user name on navbar
To dynamically render the logged in user's name on the navbar,
we use Devise's `current_user` method (I believe this method itself
is meta-programmed) to grab the user object. We also define a new
method called `name` to split the user's email and grab the first
part of their email address. Obviously this is just part of a tutorial
and is a horrible way to actually grab the name of the user. In the
future I will add a new column for the 'name' attribute.