11/28/23: render company name dynamically on the navbar
Similar to the previous commit, this commit dynamically renders the
company name on the navbar. I created a new method called `current_company`
on the ApplicationController that essentially grabs the current_company
associated with the current_user. However, I'm not a big fan of this
approach, since I feel that this logic should either live on the model,
or perhaps be extracted to an interaction. I believe that the controller
should be primarily responsible for request logic. But, for the sake of
time and following this tutorial, I'm going ahead with this change.