11/28/23: ensure signed stream names are different per company
For security reasons, I need to ensure that users from one company
cannot view the quotes from a different company, especially when
the users are subscribed to the "quotes" stream. By passing in an
array with the current_company in the array in the lambda for the
stream, we ensure that someone from a different company won't be
subscribed to the updates from a user in an another company.