11/29/23: adding flash via Turbo
This part of the tutorial adds flash support for Turbo. However,
flash notices are not popping up for me, even when I disabled
Turbo in the previous step and used regular HTML redirects.
For now, I'm going to put a hold on this step of the tutorial
because I'm not too interested in flash messages. I would like
to start incorporating openAI API calls now.