11/29/23: change up fixtures
I'm going a little off-script from the hotrails.dev tutorial since
I'm interested in incorporating the openAI API responses now. Since
I want to have openAI guess the author of a quote, I decided to change
the initial quotes to something more recognizable, like Hamlet's
"To be or not to be". That way, the user can see more easily what
exactly the openAI functionality is doing.