11/29/23: add simple button
This commit is very simple - I literally just created a new
button that doesn't direct anywhere. It's just to get the markup
on the page going. At this point, I'm thinking that I need some
kind of URL defined on a presenter object to pass to the view
that can make the call to the openAI API. Perhaps I should
use ActiveInteraction?