11/30/23: finish spike work for now
While I was able to get ActiveInteraction to render a list of
PORO versions of the ActiveRecord objects for the Quote model,
I think it got too complicated trying to update the views for the
various actions. For example, I successfully rendered the quotes
as POROs for the index page, but things got messed up when it
came to updating. So, I will conclude this spike work to utilize
ActiveInteraction, and will switch to the strategy of just getting
something to work. This probably means for my second pass I will
be passing in ActiveRecord objects directly into the view (which
I do not want to do), but I can refactor later, once I get the AI
output to actually render on the page.