11/30/23: render AI output on the screen
Wow, I was finally able to get the AI response to show on the page!
Like I mentioned in an earlier commit, this spike focuses on just
getting the functionality to work. The styling is horrible, and I'm
sure some OO principles were violated. Additionally, I'm passing in
direct ActiveRecord objects to the view, which I'm not a fan of.
However, I think it's better to duplicate first, or to get something
to work, and then you can refactor. As of now, the AI response
successfully renders, but it completely replaces the entire quote
text, so I need to figure out how to append it to the bottom of
the quote.