12/2/23: refactor edit functionality
I noticed that I was returning the entire edit partial when the #edit action was hit, even though technically I only needed to return just the form itself. So I decided to move the turbo_frame_tag to inside the form partial and surround the form itself. That way just the necessary HTML gets returned by the server.