12/3/23: use turbo stream find_author template
This time around, I'm actually able to use the turbo_stream
template instead of the turbo frame. From the UI, it's still
not quite working the way I intended. But at least I was
able to get the inner workings to operate the way I want.
Next, I will try to move that "Find Author" button outside
the turbo-frame with the quote, to see if that'll help.12/3/23: use turbo stream find_author template