12/3/23: use different id for turbo-frame
OMG IT FINALLY WORKS!! Looks like by moving the "Find Author" button
outside of the initial turbo frame and that wrapping that in its own
turbo frame worked. However, the problem is that when I update the
quote, it renders a new "Find Author" button, so I'll need to fix that.
But this is great progress for me.