12/4/23: update readme
This commit updates the README.md with the introduction, setup,
learnings, and future plans for this project. I still need to
elaborate more on the future plans, especially since I'm still
in the discovery work to incorporate Rspec for testing purposes,
and also ActiveInteraction to create a service layer for making
my API call to openAI's endpoints.