12/4/23: bug fix - don't create new buttons
Originally, when a user updated a quote, the server would create
a new "Find Author" button. I created a new turbo_stream template
for the #update action that removes the AI response for the previous
quote, and also removes the "find_author" section so that we
don't see duplicate "Find Author" buttons.

For future commits, I will work on cleaning up the styling, or
disabling the "Find Author" button while the API call is being
made. But for now, I think this PR is ok to merge.