12/5/23: disable "Find Author" button
To prevent a user from potentially clicking on the "Find Author"
button too many times, I disable the button and will replace the
button with the AI's response when it's returned. I spent a lot of
time trying to figure out how to do this via Stimulus, and it turns
out that actually setting a disabled value in the JS controller
completely blocks the 'find_author' action from being triggered.
However, adding a CSS property of 'pointer-events: none;' seemed
to do the trick, but I'm still not 100% sure if that's what actually
renders the 'disabled' attribute on the button.